% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.r
\name{make_group_without_period}
\alias{make_group_without_period}
\title{Build a grouping factor by dropping any 'period' term (e.g., f_tk)}
\usage{
make_group_without_period(fml, data, period = "f_tk", state_var)
}
\arguments{
\item{fml}{A one- or two-sided formula with fixed and/or random effects.}

\item{data}{A data.frame with all variables in the formula.}

\item{period}{Name of the time/period variable to exclude (default = "f_tk").}
}
\value{
A factor representing interaction of all non-period grouping variables.
}
\description{
Build a grouping factor by dropping any 'period' term (e.g., f_tk)
}
