% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{make_param_vectors}
\alias{make_param_vectors}
\title{Create Parameter Vectors for Process}
\usage{
make_param_vectors(design, processes = names(design$formulas))
}
\arguments{
\item{design}{Output of `make_design_list_2()`.}

\item{proc}{A string giving the process name (e.g., "phi" or "p").}
}
\value{
A list with two elements:
- `beta`: Named vector of fixed effects for this process.
- `u`: Named vector of random effects for this process (may be empty).
}
\description{
For a given process name, extract the number of fixed and random effect parameters
based on the design object and create named beta and u parameter vectors.
}
\examples{
df <- data.frame(
  f_tk   = rep(1:5, each = 6),
  f_tl   = factor(sample(c("A", "B", "C"), 30, replace = TRUE)),
  f_yr   = factor(sample(2010:2012, 30, replace = TRUE)),
  f_site = factor(sample(1:3, 30, replace = TRUE))
)

formulas <- list(
  phi = ~ f_yr + (1|f_site),
  p   = ~ 1
)

design <- make_design_list_2(formulas, list(state = "f_tl", input = "f_tk"), df)
make_param_vectors(design, "phi")
make_param_vectors(design, "p")

}
